+++ app/views/UploadInProgressView.scala.html
@@ -25,9 +25,9 @@
         cancelLink: CancelApplicationLink
 )
 
-@(draftId:DraftId, key: Option[String])(implicit request: Request[_], messages: Messages)
+@(draftId:DraftId, key: Option[String])(implicit messages: Messages, request: Request[_])
 
-@layout(pageTitle = titleNoForm(messages("uploadInProgress.title"))) {
+@layout(pageTitle = titleNoForm(messages("uploadInProgress.title")), showBackLink = false) {
 
     @formHelper(action = routes.UploadInProgressController.checkProgress(draftId, key)) {
