+++ test/controllers/UploadSupportingDocumentsControllerSpec.scala
@@ -16,19 +16,15 @@
 
 package controllers
 
-import java.time.Instant
-
 import scala.concurrent.Future
 
 import play.api.inject.bind
-import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
 import models.{NormalMode, UploadedFile}
 import models.upscan.UpscanInitiateResponse
-import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.ArgumentMatchersSugar.eqTo
 import org.mockito.Mockito.{verify, when}
@@ -38,7 +34,7 @@ import org.scalatestplus.mockito.MockitoSugar
 import pages._
 import services.UserAnswersService
 import services.fileupload.FileService
-import views.html.{UploadInProgressView, UploadSupportingDocumentsView}
+import views.html.UploadSupportingDocumentsView
 
 class UploadSupportingDocumentsControllerSpec
     extends SpecBase
@@ -70,26 +66,6 @@ class UploadSupportingDocumentsControllerSpec
 
   private val initiatedFile =
     UploadedFile.Initiated("reference")
-
-  private val successfulFile = UploadedFile.Success(
-    reference = "reference",
-    downloadUrl = "downloadUrl",
-    uploadDetails = UploadedFile.UploadDetails(
-      fileName = "fileName",
-      fileMimeType = "fileMimeType",
-      uploadTimestamp = Instant.now(),
-      checksum = "checksum",
-      size = 1337
-    )
-  )
-
-  private val failedFile = UploadedFile.Failure(
-    reference = "reference",
-    failureDetails = UploadedFile.FailureDetails(
-      failureReason = UploadedFile.FailureReason.Quarantine,
-      failureMessage = Some("failureMessage")
-    )
-  )
 
   "when there is no existing file" - {
 
@@ -180,7 +156,7 @@ class UploadSupportingDocumentsControllerSpec
 
       "when the key does not match the file" - {
 
-        "must initiate a file upload and display the page" in {
+        "must initiate a file upload and show page" in {
 
           val application = applicationBuilder(userAnswers = Some(userAnswers))
             .overrides(
@@ -255,167 +231,6 @@ class UploadSupportingDocumentsControllerSpec
           )(any())
         }
       }
-    }
-  }
-
-  "when there is a successful file" - {
-
-    val userAnswers = userAnswersAsIndividualTrader
-      .set(UploadSupportingDocumentPage, successfulFile)
-      .success
-      .value
-
-    "when the key matches the file" ignore {
-
-      "must redirect to the next page" in {
-
-        val onwardRoute = Call("GET", "/foo")
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers))
-          .overrides(
-            bind[FileService].toInstance(mockFileService),
-            bind[Navigator].to(new FakeNavigator(onwardRoute))
-          )
-          .build()
-
-        val request = FakeRequest(
-          GET,
-          controllers.routes.UploadSupportingDocumentsController
-            .onPageLoad(
-              models.NormalMode,
-              draftId,
-              None,
-              Some(successfulFile.reference)
-            )
-            .url
-        )
-
-        val result = route(application, request).value
-
-        status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual onwardRoute.url
-
-        verify(mockFileService, times(0)).initiate(any(), any(), any())(any())
-      }
-    }
-
-    "when the key does not match the file" ignore {
-
-      "must initiate a file upload and display the page" in {
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers))
-          .overrides(
-            bind[FileService].toInstance(mockFileService)
-          )
-          .build()
-
-        val view = application.injector.instanceOf[UploadSupportingDocumentsView]
-
-        when(mockFileService.initiate(any(), any(), any())(any()))
-          .thenReturn(Future.successful(upscanInitiateResponse))
-
-        val request = FakeRequest(
-          GET,
-          controllers.routes.UploadSupportingDocumentsController
-            .onPageLoad(models.NormalMode, draftId, None, Some("otherReference"))
-            .url
-        )
-
-        val result = route(application, request).value
-
-        status(result) mustEqual OK
-        contentAsString(result) mustEqual view(
-          draftId = draftId,
-          upscanInitiateResponse = Some(upscanInitiateResponse),
-          errorMessage = None
-        )(messages(application), request).toString
-
-        verify(mockFileService).initiate(
-          eqTo(draftId),
-          eqTo(redirectPath),
-          eqTo(isLetterOfAuthority)
-        )(any())
-      }
-    }
-
-    "when there is no key" ignore {
-
-      "must initiate a file upload and display the page" in {
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers))
-          .overrides(
-            bind[FileService].toInstance(mockFileService)
-          )
-          .build()
-
-        val view = application.injector.instanceOf[UploadSupportingDocumentsView]
-
-        when(mockFileService.initiate(any(), any(), any())(any()))
-          .thenReturn(Future.successful(upscanInitiateResponse))
-
-        val request = FakeRequest(
-          GET,
-          controllers.routes.UploadSupportingDocumentsController
-            .onPageLoad(models.NormalMode, draftId, None, None)
-            .url
-        )
-
-        val result = route(application, request).value
-
-        status(result) mustEqual OK
-        contentAsString(result) mustEqual view(
-          draftId = draftId,
-          upscanInitiateResponse = Some(upscanInitiateResponse),
-          errorMessage = None
-        )(messages(application), request).toString
-
-        verify(mockFileService).initiate(
-          eqTo(draftId),
-          eqTo(redirectPath),
-          eqTo(isLetterOfAuthority)
-        )(any())
-      }
-    }
-  }
-
-  "when there is a failed file" ignore {
-
-    val userAnswers =
-      userAnswersAsIndividualTrader
-        .set(UploadSupportingDocumentPage, failedFile)
-        .success
-        .value
-
-    "must initiate a file upload and redirect back to the page with the relevant error code" in {
-
-      val application = applicationBuilder(userAnswers = Some(userAnswers))
-        .overrides(
-          bind[FileService].toInstance(mockFileService)
-        )
-        .build()
-
-      when(mockFileService.initiate(any(), any(), any())(any()))
-        .thenReturn(Future.successful(upscanInitiateResponse))
-
-      val request = FakeRequest(
-        GET,
-        controllers.routes.UploadSupportingDocumentsController
-          .onPageLoad(models.NormalMode, draftId, None, Some("key"))
-          .url
-      )
-
-      val result = route(application, request).value
-
-      status(result) mustEqual SEE_OTHER
-      redirectLocation(result).value mustEqual routes.UploadSupportingDocumentsController
-        .onPageLoad(models.NormalMode, draftId, Some("Quarantine"), Some("key"))
-        .url
-
-      verify(mockFileService).initiate(
-        eqTo(draftId),
-        eqTo(redirectPath),
-        eqTo(isLetterOfAuthority)
-      )(any())
     }
   }
